refactor!: removed OpenAI and Claude Summarizers
Also refactored and simplified config and main file. From now is not needed to specify summarizer type because, program uses only Ollama

BREAKING CHANGE: OpenAI and Claude are no avaialbale for summarizing anymore. the type field in config is removed
