+++ nitrodigest/config.py
@@ -2,13 +2,6 @@ import json
 import os
 from typing import Optional, Dict, Any
 from dataclasses import dataclass
-from enum import Enum
-
-
-class SummarizerType(Enum):
-    CLAUDE = "claude"
-    CHATGPT = "chatgpt"
-    OLLAMA = "ollama"
 
 
 @dataclass
@@ -28,19 +21,14 @@ class EmailConfig:
 
 @dataclass
 class SummarizerConfig:
-    type: SummarizerType
-    model: Optional[str] = None
+    model: Optional[str] = 'mistral'
     base_url: str = "http://localhost:11434"
-    api_key: Optional[str] = None
     timeout: int = 300
-    prompt_file: Optional[str] = None
+    prompt_file: Optional[str] = 'prompt_template.txt'
 
     def validate(self) -> None:
-        if self.type == SummarizerType.OLLAMA and not self.model:
+        if not self.model:
             raise ValueError("Model is required for Ollama")
-        if (self.type in [SummarizerType.CLAUDE, SummarizerType.CHATGPT] and
-                not self.api_key):
-            raise ValueError("API key is required for Claude and ChatGPT")
         if self.timeout <= 0:
             raise ValueError("Timeout must be a positive number")
         if (self.prompt_file and
@@ -70,20 +58,6 @@ class Config:
         if not self.email.folder:
             raise ValueError("Email folder is required")
 
-        if not self.summarizer.type:
-            raise ValueError("Summarizer type is required")
-        if (self.summarizer.type in [SummarizerType.CLAUDE,
-                                     SummarizerType.CHATGPT] and
-                not self.summarizer.api_key):
-            raise ValueError("API key is required for Claude and ChatGPT")
-        if (self.summarizer.type == SummarizerType.OLLAMA and
-                not self.summarizer.model):
-            raise ValueError("Model is required for Ollama")
-        if (self.summarizer.prompt_file and
-                not os.path.exists(self.summarizer.prompt_file)):
-            raise ValueError(
-                f"Prompt file not found: {self.summarizer.prompt_file}")
-
         self.email.validate()
         self.summarizer.validate()
 
@@ -99,23 +73,16 @@ class Config:
         )
 
         summarizer_data = data['summarizer']
-        summarizer_type = SummarizerType(summarizer_data['type'])
         model = summarizer_data.get('model')
         if model is None:
-            if summarizer_type == SummarizerType.CLAUDE:
-                model = "claude-3-haiku-20240307"
-            elif summarizer_type == SummarizerType.CHATGPT:
-                model = "gpt-3.5-turbo"
-            elif summarizer_type == SummarizerType.OLLAMA:
-                model = "mistral"
+            model = "mistral"
 
         summarizer_config = SummarizerConfig(
-            type=summarizer_type,
             model=model,
             base_url=summarizer_data.get('base_url', 'http://localhost:11434'),
-            api_key=summarizer_data.get('api_key'),
             timeout=summarizer_data.get('timeout', 300),
-            prompt_file=summarizer_data.get('prompt_file')
+            prompt_file=summarizer_data.get(
+                'prompt_file', 'prompt_template.txt'),
         )
 
         return cls(
@@ -144,7 +111,6 @@ class Config:
                 'folder': self.email.folder
             },
             'summarizer': {
-                'type': self.summarizer.type.value,
                 'model': self.summarizer.model,
                 'base_url': self.summarizer.base_url,
                 'timeout': self.summarizer.timeout
@@ -154,8 +120,6 @@ class Config:
             'mark_as_read': self.mark_as_read
         }
 
-        if self.summarizer.api_key:
-            data['summarizer']['api_key'] = self.summarizer.api_key
         if self.summarizer.prompt_file:
             data['summarizer']['prompt_file'] = self.summarizer.prompt_file
