--- nitrodigest/summarizer/providers/__init__.py
@@ -1,13 +1,0 @@
-"""
-Provider-specific summarizer implementations.
-"""
-
-from summarizer.providers.claude import ClaudeSummarizer
-from summarizer.providers.openai import ChatGPTSummarizer
-from summarizer.providers.ollama import OllamaSummarizer
-
-__all__ = [
-    'ClaudeSummarizer',
-    'ChatGPTSummarizer',
-    'OllamaSummarizer'
-]
